NEX-15: Reorganize component directories for article and forms components
