+++ next/components/article/articles-listing.tsx
@@ -2,7 +2,7 @@ import { useRouter } from "next/router";
 import { useTranslation } from "next-i18next";
 import { useQuery } from "@tanstack/react-query";
 
-import { ArticleTeaser } from "@/components/article-teaser";
+import { ArticleTeaser } from "@/components/article/article-teaser";
 import { LoadingSpinner } from "@/components/loading-spinner";
 import type { ArticleTeaserType } from "@/types/graphql";
