+++ next/pages/all-articles/[[...page]].tsx
@@ -2,7 +2,7 @@ import { GetStaticPaths, GetStaticProps, InferGetStaticPropsType } from "next";
 import { useTranslation } from "next-i18next";
 import { useRef } from "react";
 
-import { ArticleListItem } from "@/components/article-list-item";
+import { ArticleListItem } from "@/components/article/article-list-item";
 import { HeadingPage } from "@/components/heading--page";
 import { LayoutProps } from "@/components/layout";
 import { Meta } from "@/components/meta";
