+++ next/pages/auth/login.tsx
@@ -6,7 +6,7 @@ import { useTranslation } from "next-i18next";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
 
-import { ErrorRequired } from "@/components/error-required";
+import { ErrorRequired } from "@/components/forms/error-required";
 import { Meta } from "@/components/meta";
 import { getCommonPageProps } from "@/lib/get-common-page-props";
 import { useEffectOnce } from "@/lib/hooks/use-effect-once";
