+++ next/pages/auth/register.tsx
@@ -3,7 +3,7 @@ import { useRouter } from "next/router";
 import { useTranslation } from "next-i18next";
 import { useForm } from "react-hook-form";
 
-import { ErrorRequired } from "@/components/error-required";
+import { ErrorRequired } from "@/components/forms/error-required";
 import { Meta } from "@/components/meta";
 import { getCommonPageProps } from "@/lib/get-common-page-props";
